Fixes some issues with regions that were causing DFTF to crash: regions were scaled wrong when saved to registry (config memory - different from OLN files) such that they would typically go way off the left edge of the igram and not be visible to users.
Also fixed a bug where regions would be changed to their encompasing square.  But only in debug mode so didn't affect users.

Also changed it so regions are saved to registry (config) when you hit "done" and
not when you hit "crop".  This way if the user makes a mistake with a region
or other outline they can just reload the igram.